make keyboa_run argument list consistent
keyboa_run used to just reduce the list of transformation by function
application. This meant that all items could be generators except the
last item that had to be a function. This commit removes that
exception so that all items in the list can be, and are required to
be, transformations.

Also fix output to be a transformation rather than a function.